+++ Code/resource_view.h
@@ -28,7 +28,7 @@ class ResourceView {
     //SafeRelease(&data_pointer);
   }
   void* data_pointer;
-  uint32_t data_length;
+  size_t data_length;
 };
 
 }
